Removed console output / moved files
